Extract some classes to their own source files
Also remove more unused dto classes.

No changes semantically
