+++ src/main/scala/cognite/spark/v1/CdfPartition.scala
@@ -1,0 +1,5 @@
+package cognite.spark.v1
+
+import org.apache.spark.Partition
+
+final case class CdfPartition(index: Int) extends Partition
