+++ src/main/scala/cognite/spark/v1/DefaultSource.scala
@@ -23,65 +23,6 @@ import org.apache.spark.sql.{DataFrame, Row, SQLContext, SaveMode}
 import sttp.client3.SttpBackend
 import sttp.model.Uri
 
-final case class RelationConfig(
-    auth: CdfSparkAuth,
-    clientTag: Option[String],
-    applicationName: Option[String],
-    projectName: String,
-    batchSize: Option[Int],
-    limitPerPartition: Option[Int],
-    partitions: Int,
-    maxRetries: Int,
-    maxRetryDelaySeconds: Int,
-    collectMetrics: Boolean,
-    collectTestMetrics: Boolean,
-    metricsPrefix: String,
-    baseUrl: String,
-    onConflict: OnConflictOption,
-    applicationId: String,
-    parallelismPerPartition: Int,
-    ignoreUnknownIds: Boolean,
-    deleteMissingAssets: Boolean,
-    subtrees: AssetSubtreeOption,
-    ignoreNullFields: Boolean,
-    rawEnsureParent: Boolean
-) {
-
-  /** Desired number of Spark partitions ~= partitions / parallelismPerPartition */
-  def sparkPartitions: Int = Math.max(1, partitions / parallelismPerPartition)
-}
-
-sealed trait OnConflictOption extends Serializable
-object OnConflictOption {
-  object Abort extends OnConflictOption
-  object Update extends OnConflictOption
-  object Upsert extends OnConflictOption
-  object Delete extends OnConflictOption
-  val fromString: Map[String, OnConflictOption] = Map(
-    "abort" -> Abort,
-    "update" -> Update,
-    "upsert" -> Upsert,
-    "delete" -> Delete
-  )
-
-  def withNameOpt(s: String): Option[OnConflictOption] =
-    fromString.get(s.toLowerCase)
-}
-
-sealed trait AssetSubtreeOption extends Serializable
-object AssetSubtreeOption {
-  object Ingest extends AssetSubtreeOption
-  object Ignore extends AssetSubtreeOption
-  object Error extends AssetSubtreeOption
-  val fromString: Map[String, AssetSubtreeOption] = Map(
-    "ingest" -> Ingest,
-    "ignore" -> Ignore,
-    "error" -> Error
-  )
-  def withNameOpt(s: String): Option[AssetSubtreeOption] =
-    fromString.get(s.toLowerCase)
-}
-
 class DefaultSource
     extends RelationProvider
     with CreatableRelationProvider
